reduce scope of 'Ferm::Actions'
If 'Ferm::Actions' is equal to some defined strings or any non-empty
string it's almost useless. For this reason the defined type 'rule' has
to reiterate testing against the same Enum to get useful information
from it.

'Ferm::Actions' is used in 'rule' and 'ipset'. The later does not seem
to require jumping.

update spec/type_aliases/actions_spec.rb